pyproject.toml
[tool.poetry]
name = "jazzbuck-automation"
version = "0.1.0"
description = "Scripts to run on a web server to help me in my life."
authors = ["jazzbuck"]
license = "MIT"
readme = "README.md"
packages = [{include = "jazzbuck_automation"}]

[tool.poetry.dependencies]
python = "^3.10"
orjson = "^3.8.5"
requests = "^2.28.2"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
black = "^22.12.0"
pytest-black = "^0.3.12"
flake8 = "^6.0.0"
pre-commit = "^2.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
